backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pymongo[srv]==4.6.1
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.2
email-validator==2.1.0
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
openpyxl==3.1.2
prophet==1.1.5
joblib>=1.3.0

# CSP SDKs (install only the ones you need)
# Azure
# azure-identity>=1.15.0
# azure-mgmt-costmanagement>=4.0.0

# AWS
# boto3>=1.34.0

# GCP
# google-cloud-bigquery>=3.14.0
